% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aalborg-1855-census.R
\docType{data}
\name{aalborg_1855_census}
\alias{aalborg_1855_census}
\title{Aalborg 1855 census: age & gender of every individual}
\format{A data frame with 7772 rows and 2 variables:
\describe{
  \item{gender}{character vector where M = male and F = female}
  \item{age}{Numeric vector of age of individual as recorded in census}
  }}
\source{
{Aalborg census 1855. Data entered by danishfamilysearch.com}
}
\usage{
aalborg_1855_census
}
\description{
A dataset containing the age and gender of every individual recorded in 1850
census.
}
\keyword{datasets}

