% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aalborg-age-gender.R
\docType{data}
\name{aalborg_age_gender}
\alias{aalborg_age_gender}
\title{Cholera in Aalborg by age and gender}
\format{A data frame with 10 rows and 3 variables:
\describe{
  \item{\bold{age_grop}}{Character vector of the age grouping}
  \item{\bold{male_sick}}{Integer vector of the number of males diagnosed
  with cholera}
  \item{\bold{male_dead}}{Integer vector of the number of males recorded as
  dying due to cholera}
  \item{\bold{female_sick}}{Integer vector of the number of females diagnosed
  with cholera}
  \item{\bold{female_dead}}{Integer vector of the number of females recorded as
  dying due to cholera}
  \item{\bold{total_sick}}{Integer vector of the total number of people
  diagnosed with cholera}
  \item{\bold{total_dead}}{Integer vector of the total number of people recorded as
  dying due to cholera}
  }}
\source{
{Mads to tell}
}
\usage{
aalborg_age_gender
}
\description{
A dataset containing the number of cholera cases and deaths broken down by
age and gender in Aalborg in 1853. Data comes from...
}
\keyword{datasets}
