% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_cause_1853.R
\docType{data}
\name{all_cause_1852_1854}
\alias{all_cause_1852_1854}
\title{Copenhagen mortality due to diarrhea, cholera and all-cause in 1853}
\format{A data frame with 36 rows and 5 variables:
\describe{
  \item{date}{the month and year}
  \item{cholera}{Number of dead due to cholera}
  \item{diarrhea}{Number dead due to diarrhea}
  \item{all}{Number dead due to all causes, include cholera and diarrhea}
  }}
\source{
{}
}
\usage{
all_cause_1852_1854
}
\description{
A dataset of monthly mortality for 1852 to 1854 due to cholera, diarrhea, and
all cause
}
\keyword{datasets}
