% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/broad_st.R
\docType{data}
\name{broad_st}
\alias{broad_st}
\title{Daily mortality counts in the London Broad Street outbreak}
\format{A data frame with 7 rows and 3 variables:
\describe{
  \item{\bold{fatal_attacks}}{Unknown}
  \item{\bold{deaths}}{unknown}
  \item{\bold{week}}{week number of epidemic}
  \item{\bold{date}}{date}
  }}
\source{
{On the mode and communication of Cholera by John Snow}
}
\usage{
broad_st
}
\description{
A dataset containing the number of deaths each day as tabulated by John Snow during the Broad Street pump outbreak in Soho, London, 1854
}
\keyword{datasets}

