% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cph_age_pop1853_raw.R
\docType{data}
\name{cph_age_pop1853_raw}
\alias{cph_age_pop1853_raw}
\title{Copenhagen population by age group 1853}
\format{A data frame with 23 rows and 4 variables:
\describe{
  \item{age_range}{The age range}
  \item{men1853}{Numeric vector of the number of men in the age group}
  \item{women1853}{Numeric vector of the number of women in the age group}
  \item{total1853}{Numeric vector of the total number of men & women in the
  age group}
  }}
\source{
{Copenhagen census 1850 & 1853. Data entered by ___}
}
\usage{
cph_age_pop1853_raw
}
\description{
A dataset of the population, by gender, in each age group. Data is imputed
from the 1850 & 1855 census. The age groupings used are the same groups used
in both censuses
}
\keyword{datasets}

