% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cph_age_sick_dead_raw.R
\docType{data}
\name{cph_age_sick_dead_raw}
\alias{cph_age_sick_dead_raw}
\title{Copenhagen morbidity and mortality numbers from cholera, 1853}
\format{A data frame with 23 rows and 4 variables:
\describe{
  \item{age_range}{The age range}
  \item{male_sick}{Numeric vector of the number of men recorded with cholera}
  \item{female_sick}{Numeric vector of the number of women recorded with
  cholera}
  \item{total_sick}{Numeric vector of the total number of men & women recorded with cholera}
  \item{men_dead}{Numeric vector of the number of men who died due to cholera}
  \item{women_dead}{Numeric vector of the number of women who died due to cholera}
  \item{total_dead}{Numeric vector of the total number of men & women who died due to cholera}
  }}
\source{
{Beretning om cholera epidemien i Kjöbenhavn. Data entered by ___}
}
\usage{
cph_age_sick_dead_raw
}
\description{
A dataset of the total morbidity and mortality counts due to cholera for the
age groups deliniated by the census
}
\keyword{datasets}
