% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPH-monthly-all-cause-mortality_desc.R
\docType{data}
\name{CPH_allcause}
\alias{CPH_allcause}
\title{Copenhagen mortality by age group 1853}
\format{A data frame with 36 rows and 5 variables:
\describe{
  \item{date}{the month and year}
  \item{<10}{All cause mortality in people under 10}
  \item{10-25}{All cause mortality in people aged 10-25 }
  \item{26-50}{All cause mortality in people aged 26-50}
  \item{>50}{All cause mortality in people older than 50}
  }}
\source{
{}
}
\usage{
CPH_allcause
}
\description{
A dataset of monthly all-cause mortality broken into four age-groups Copenhagen
for the years of 1852 - 1854
}
\keyword{datasets}
