% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cph_quarter_data.R, R/cph_street_data.R
\docType{data}
\name{cph_street_data}
\alias{cph_street_data}
\title{Cholera cases by week by quarter - Copenhagen 1853.}
\format{A data frame with 416 rows and 17 variables:
\describe{
  \item{quarter}{Name of the quarter (known in American english as
  "neigborhood"), that street is located in. Hospitals are given their own
   quarter name, not the name of the quarter that the hospital is
   physically located in.}
  \item{start.date}{Date of the beginig of the week that the row pertains to.}
  \item{startday.index}{The day number, as indexed fromt he beginig of the
  epidemic, of the week that the row pertains to.}
  \item{mensick.week}{Number of males recorded with cholera in the quarter
  during the specified week.}
  \item{mendead.week}{Of those men who were sick, this is the number that died
  due to cholera during the specified week.}
  \item{womensick.week}{Number of females recorded with cholera in the quarter
  during the specified week.}
  \item{womendead.week}{Of those women who were sick, this is the number that died
  due to cholera during the specified week.}
  \item{sick.total.week}{Total number of people recorded with cholera in the quarter
  during the specified week.}
  \item{dead.total.week}{Total number of people who died due to cholera in the quarter
  during the specified week.}
  \item{est.pop.1853}{Estimated population for the quarter. Estimate made by
  estimating yearly population change between 1850 and 1855 and adding 3 years
  of this to the 1850 census data for the quarter}
  \item{normal.incidence}{Weekly incidence rate per 1000 people for each quarter normalized using
 est.pop.1853}
  \item{normal.mortality}{Weekly mortality rate per 1000 people for each
 quarter normalized using est.pop.1853 population data}
  \item{est.pop.1853}{}
  \item{est.pop.1853}{}
  \item{est.pop.1853}{}
  \item{est.pop.1853}{}
  \item{est.pop.1853}{}
  \item{est.pop.1853}{}
}}
\source{
{"Beretning om Cholera-Epidemien I Kjobenhavn"}

{"Beretning om Cholera-Epidemien I Kjobenhavn"}
}
\usage{
cph_street_data

cph_street_data
}
\description{
A dataset of the weekly record of new cholera cases in each quarter (also
called neighborhoods) in Copenhagen in the 1853 epidemic. Dataset also 
includs hospitals and quarters from outside the city walls (i.e. Noerrebro).
Danish letters have been turned into Engilsh letters: oe, ae, aa.
Data was manually entered by author from source report. Please contact
author with any errors or typos

A dataset of the weekly record of new cholera cases on each street
in Copenhagen in the 1853 epidemic. Dataset also includs hospitals
and streets from outside the city walls (i.e. Noerrebro). Danish letters
have been turned into Engilsh letters: oe, ae, aa. Data was manually entered
by author from source report. Please contact author with any errors or typos
}
\keyword{datasets}

