% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DK-cities-monthly-all-cause-mortality_desc.R
\docType{data}
\name{dk_cities_allcause}
\alias{dk_cities_allcause}
\title{Copenhagen population by age group 1853}
\format{A data frame with 23 rows and 4 variables:
\describe{
  \item{date}{the month}
  \item{<10}{All cause mortality in people under 10}
  \item{10-25}{All cause mortality in people aged 10-25 }
  \item{26-50}{All cause mortality in people aged 26-50}
  \item{>50}{All cause mortality in people older than 50}
  }}
\source{
{Copenhagen census 1850 & 1853. Data entered by ___}
}
\usage{
dk_cities_allcause
}
\description{
A dataset of monthly all-cause mortality broken into 4 age groups in Danish
cities (not including Copenhagen) for the years of 1852 - 1854
}
\keyword{datasets}
