% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/korsoer-age-gender-pop.R
\docType{data}
\name{korsoer_age_gender_pop}
\alias{korsoer_age_gender_pop}
\title{Population in Korsoer by age and gender}
\format{A data frame with 9 rows and 5 variables:
\describe{
  \item{\bold{age_grop}}{Character vector of the age grouping}
  \item{\bold{f}}{Integer of number of females in each age group}
  \item{\bold{m}}{Integer of number of males in each age group}
  \item{\bold{u}}{Integer of number of people of unknown gender in each age group}
  \item{\bold{Total}}{Integer of number of males, females and unkowns in each
  age group} }}
\source{
{Ask Lone}
}
\usage{
korsoer_age_gender_pop
}
\description{
A dataset containing the number of people broken down by
age and gender in Korsoer in 1853. Data comes from...
}
\keyword{datasets}
